Renames lib/precompiled-abi-forks-util.js to lib/bridge-provider.js
