+++ lib/2wp-utils.js
@@ -10,7 +10,7 @@ const {
 const BN = require('bn.js');
 const { retryWithCheck, ensure0x } = require('./utils');
 const { waitForBitcoinTxToBeInMempool, waitForBitcoinMempoolToGetTxs, getBtcAddressBalanceInSatoshis } = require('./btc-utils');
-const { getBridge } = require('./precompiled-abi-forks-util');
+const { getBridge } = require('./bridge-provider');
 const { getBridgeState } = require('@rsksmart/bridge-state-data-parser');
 const { getDerivedRSKAddressInformation } = require('@rsksmart/btc-rsk-derivation');
 const btcEthUnitConverter = require('@rsksmart/btc-eth-unit-converter');
