+++ lib/assertions/whitelisting.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect;
 const rskUtils = require('../rsk-utils');
-const { getBridge } = require('../precompiled-abi-forks-util');
+const { getBridge } = require('../bridge-provider');
 
 const WHITELIST_CHANGE_PK = '3890187a3071327cee08467ba1b44ed4c13adb2da0d5ffcc0563c371fa88259c';
 const WHITELIST_CHANGE_ADDR = '87d2a0f33744929da08b65fd62b627ea52b25f8e';
