+++ lib/bridge-provider.js
