+++ lib/rsk-utils.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect;
 const { getBridgeState } = require('@rsksmart/bridge-state-data-parser');
-const { getBridge } = require('./precompiled-abi-forks-util');
+const { getBridge } = require('./bridge-provider');
 const BridgeTransactionParser = require('@rsksmart/bridge-transaction-parser');
 const { getRskTransactionHelpers } = require('../lib/rsk-tx-helper-provider');
 const { wait, retryWithCheck, removePrefix0x } = require('./utils');
