+++ lib/tests/2wp.js
@@ -1,7 +1,7 @@
 const expect = require('chai').expect;
 const BN = require('bn.js');
 const { createPeginV1TxData } = require('pegin-address-verificator');
-const { getBridge } = require('../precompiled-abi-forks-util');
+const { getBridge } = require('../bridge-provider');
 const { getBtcClient } = require('../btc-client-provider');
 const { getRskTransactionHelper, getRskTransactionHelpers } = require('../rsk-tx-helper-provider');
 const { satoshisToBtc, btcToSatoshis, satoshisToWeis } = require('@rsksmart/btc-eth-unit-converter');
